+++ parser/src/price/mod.rs
@@ -1,5 +1,7 @@
 mod binance;
 mod kraken;
+mod satonomics;
 
 pub use binance::*;
 pub use kraken::*;
+pub use satonomics::*;